change command tags type form int32_t to uint64_t

+++ libscpi/inc/scpi/parser.h
@@ -128,7 +128,7 @@ extern "C" {
 
     scpi_bool_t SCPI_IsCmd(scpi_t * context, const char * cmd);
 #if USE_COMMAND_TAGS
-    int32_t SCPI_CmdTag(scpi_t * context);
+    uint64_t SCPI_CmdTag(scpi_t * context);
 #endif /* USE_COMMAND_TAGS */
     scpi_bool_t SCPI_Match(const char * pattern, const char * value, size_t len);
     scpi_bool_t SCPI_CommandNumbers(scpi_t * context, int32_t * numbers, size_t len, int32_t default_value);
